Add arguments to ExampleServiceHost to add users
This add arguments to ExampleServiceHost to create two users: an
administrative user with access to everything and a user that only has
access to ExampleService documents that they own. This is meant for
illustrative purposes: other service hosts will likely make users in a
more complex roles.

Findbugs reports spurious warnings on the ExampleServiceHost
arguments, so I added exclusions for them.

Coverage results look good.

Change-Id: Ia320729646fccb56d280dd0cef5da3c09c1f2c6d