Enforce NodeState.synchQuorum
Automatic synchronization will only be triggered when in addition to
stable membership (no new updates from gossip), the number of AVAILABLE
nodes is more or equal to the nodeState.synchQuorum. This means that
if a node group that started with 5 nodes, and syncQuorum of 5, loses a node,
any synchronization/owner selection at the factory level will only trigger
when 2 new nodes join to bring membership back to 5.

The node group IS available however, updates, GETs will be served and routed

This is all relevant only when automatic synchronization is enabled.

https://www.pivotaltracker.com/story/show/111275624

Change-Id: I88b5b213fe1921194ad81dff533b97ef52646def