Added ServiceOption.CONCURRENT_GET_HANDLING
Service runtime allows a GET to execute concurrently with updates.
This option is enabled by default. Disabling this option helps in situations where,
for example, an operation processing filter reads the current state and then
conditionally updates it, relying on the state not being updated by some other
operation in the meantime.

Change-Id: I4284f83ecd3e472a14e0354755644a41a65cbed1