Implement authorization of stateless services
New method authorizeRequest() added to Service interface.

Service classes that extend StatelessService must now be
explicitly authorized (by adding their self link in the
resource specification of a role). If a stateless service
wants to bypass authorization it must override authorizeRequest()

Factory services bypass authorization since we apply authZ to
the POST body, and we also apply it on the GET response. If a
author wants stricter authz on a factory they can implement
authorizeRequest() and call ServiceHost.isAuthorized(), which is
now public.

https://www.pivotaltracker.com/story/show/108049106

Change-Id: I39c1b262bca59ac01c36c4b4370357c70f98b6e3