Fail build if Findbugs finds any problems.
This could have been done by flipping a flag, but I improved how we
use Findbugs. Specifically:

- Every time Maven compiles it runs Findbugs and fails if Findbugs
  fails. This is good because it means we don't need to use the verify
  phase, but we just always check Findbugs problems. It does make
  compilation a bit slower.
- Unified configuration for findbugs during compilation and reporting
- Told Eclipse not to run Findbugs during compilation, for fast
  compilation cycles in Eclipse.

Note that the Jenkin's dcp build will still run Findbugs even though
we moved it to from the verify phase to the compile file.

Tip of the day: If you get Findbugs warnings, it's a pain to read the
XML and kind of annoying to generate HTML. You can do a command like
this to see the Findbugs problems you got: (If you don't include the
--projects flag, it will run it once for each Mavn project.)

    mvn --projects xenon-common findbugs:gui

Change-Id: I5a6629585b59dc6a0a784736dbca4cb7e65dc926