Implement /available utility suffix
Provide a standard mechanism to declare a service ready
for requests. The default behavior is that <service>/available
will return 200, if the service has started.

A service author can explicitly control this by calling the
new helper method setAvailable(true|false).

Since the mechanism uses a service stat under the covers, it
re-uses the stats reporting in xenon (per service).

The availability can be set either in code, within the service
or using the REST api (GET /service/available or PUT /service/available)

https://github.com/vmware/xenon/wiki/REST-API#utility-services-per-instance

https://www.pivotaltracker.com/story/show/112962821

Change-Id: Ia02f1f5701c9ed82dc140a5c3c538f924fa8cadd