Introduce PRAGMA_FORCE_INDEX_UPDATE
This can be used by the runtime to force updates on conflicting
state. It can also be used by clients that want to re-create
a service that was previously deleted, but without having to query
the DELETE version (so they can add one to it)

Extended tests to verify behavior.

Also improved the service availability check and changed the
verification host to use the new, synch aware, /available.

Change-Id: If93790fa1f4e63991fc76c3f98dfc4639d61f7a5