Use a minimum maintenance interval of 1 millisecond
Make sure maintenance interval of ServiceHost, StatelessService
and StatefulService can't be set to a value less of 1 millisecond.

This is to resolve an issue of not calling the service periodic maintenance when
the service maintenance interval value is less than <8 microsecond. The deadline time
in performMaintenanceStage is calculated using the host maintenance interval value,
and will be too low to have enough time to process services in the SERVICE stage.

Change-Id: I1143ca4fa55bb9fadded6a8e84df26573d4023bd