Fix default failure completion log message
This is not a critical issue since this affects operations
that did not even have a completion to begin with.

In this case, our logs assumed context id was a number, which is
weird since we do not assume this, with 100% certainty, anywhere else.
I believe, when this was added, we wanted to print the operation ID
which is a long, NOT the context id ...

https://www.pivotaltracker.com/story/show/115184841

Change-Id: I6172d39d073b3fe778c77538a6e6bfcb29964374